+++ saleor/core/jwt_manager.py
@@ -7,11 +7,15 @@ import jwt
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 from django.conf import settings
+from django.contrib.sites.models import Site
 from django.core.exceptions import ImproperlyConfigured
 from django.core.management.color import color_style
+from django.urls import reverse
 from django.utils.module_loading import import_string
 from jwt import api_jws
 from jwt.algorithms import RSAAlgorithm
+
+from .utils import build_absolute_uri
 
 logger = logging.getLogger(__name__)
 
@@ -20,6 +24,10 @@ KID = "1"
 
 
 class JWTManagerBase:
+    @classmethod
+    def get_domain(cls) -> str:
+        return NotImplemented
+
     @classmethod
     def get_private_key(cls) -> rsa.RSAPrivateKey:
         return NotImplemented
@@ -37,7 +45,9 @@ class JWTManagerBase:
         return NotImplemented
 
     @classmethod
-    def decode(cls, token: str, verify_expiration: bool = True) -> dict:
+    def decode(
+        cls, token: str, verify_expiration: bool = True, verify_aud: bool = False
+    ) -> dict:
         return NotImplemented
 
     @classmethod
@@ -48,9 +58,17 @@ class JWTManagerBase:
     def get_jwks(cls) -> dict:
         return NotImplemented
 
+    @classmethod
+    def get_issuer(cls) -> str:
+        return NotImplemented
+
 
 class JWTManager(JWTManagerBase):
     KEY_FILE_FOR_DEBUG = ".jwt_key.pem"
+
+    @classmethod
+    def get_domain(cls) -> str:
+        return Site.objects.get_current().domain
 
     @classmethod
     def get_private_key(cls) -> rsa.RSAPrivateKey:
@@ -134,20 +152,22 @@ class JWTManager(JWTManagerBase):
         )
 
     @classmethod
-    def decode(cls, token, verify_expiration: bool = True):
+    def decode(cls, token, verify_expiration: bool = True, verify_aud: bool = False):
+        # `verify_aud` set to false as we decode our own tokens
+        # we can have `aud` defined for app or custom.
         headers = jwt.get_unverified_header(token)
         if headers.get("alg") == "RS256":
             return jwt.decode(
                 token,
                 cls.get_public_key(),  # type: ignore
                 algorithms=["RS256"],
-                options={"verify_exp": verify_expiration},
+                options={"verify_exp": verify_expiration, "verify_aud": verify_aud},
             )
         return jwt.decode(
             token,
             settings.SECRET_KEY,  # type: ignore
             algorithms=["HS256"],
-            options={"verify_exp": verify_expiration},
+            options={"verify_exp": verify_expiration, "verify_aud": verify_aud},
         )
 
     @classmethod
@@ -169,6 +189,10 @@ class JWTManager(JWTManagerBase):
             cls.get_private_key()
         except Exception as e:
             raise ImproperlyConfigured(f"Unable to load provided PEM private key. {e}")
+
+    @classmethod
+    def get_issuer(cls) -> str:
+        return build_absolute_uri(reverse("api"), domain=cls.get_domain())
 
 
 def get_jwt_manager() -> JWTManagerBase: