+++ saleor/graphql/checkout/mutations/checkout_customer_attach.py
@@ -74,19 +74,19 @@ class CheckoutCustomerAttach(BaseMutation):
             )
 
         user_id_from_request = None
-        if not info.context.user.is_anonymous:
-            user_id_from_request = graphene.Node.to_global_id(
-                "User", info.context.user.id
-            )
+        if user := info.context.user:
+            user_id_from_request = graphene.Node.to_global_id("User", user.id)
 
         if customer_id and customer_id != user_id_from_request:
             requestor = get_user_or_app_from_context(info.context)
-            if not requestor.has_perm(AccountPermissions.IMPERSONATE_USER):
+            if not requestor or not requestor.has_perm(
+                AccountPermissions.IMPERSONATE_USER
+            ):
                 raise PermissionDenied(
                     permissions=[AccountPermissions.IMPERSONATE_USER]
                 )
             customer = cls.get_node_or_error(info, customer_id, only_type="User")
-        elif info.context.user.is_anonymous:
+        elif not info.context.user:
             raise ValidationError(
                 {
                     "customer_id": ValidationError(