+++ saleor/order/tests/test_order.py
@@ -94,7 +94,6 @@ def test_add_variant_to_order_adds_line_for_new_variant(
     product,
     product_translation_fr,
     settings,
-    anonymous_user,
     anonymous_plugins,
     site_settings,
 ):
@@ -107,7 +106,7 @@ def test_add_variant_to_order_adds_line_for_new_variant(
     add_variant_to_order(
         order=order,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,
@@ -134,7 +133,6 @@ def test_add_variant_to_order_adds_line_for_new_variant_on_sale(
     sale,
     discount_info,
     settings,
-    anonymous_user,
     anonymous_plugins,
     site_settings,
 ):
@@ -149,7 +147,7 @@ def test_add_variant_to_order_adds_line_for_new_variant_on_sale(
     add_variant_to_order(
         order=order,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,
@@ -182,7 +180,6 @@ def test_add_variant_to_draft_order_adds_line_for_variant_with_price_0(
     product,
     product_translation_fr,
     settings,
-    anonymous_user,
     anonymous_plugins,
     site_settings,
 ):
@@ -199,7 +196,7 @@ def test_add_variant_to_draft_order_adds_line_for_variant_with_price_0(
     add_variant_to_order(
         order=order,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,
@@ -218,7 +215,6 @@ def test_add_variant_to_draft_order_adds_line_for_variant_with_price_0(
 def test_add_variant_to_order_not_allocates_stock_for_new_variant(
     order_with_lines,
     product,
-    anonymous_user,
     anonymous_plugins,
     site_settings,
 ):
@@ -231,7 +227,7 @@ def test_add_variant_to_order_not_allocates_stock_for_new_variant(
     add_variant_to_order(
         order=order_with_lines,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,
@@ -242,7 +238,7 @@ def test_add_variant_to_order_not_allocates_stock_for_new_variant(
 
 
 def test_add_variant_to_order_edits_line_for_existing_variant(
-    order_with_lines, anonymous_user, anonymous_plugins, site_settings
+    order_with_lines, anonymous_plugins, site_settings
 ):
     existing_line = order_with_lines.lines.first()
     variant = existing_line.variant
@@ -255,7 +251,7 @@ def test_add_variant_to_order_edits_line_for_existing_variant(
     add_variant_to_order(
         order=order_with_lines,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,
@@ -269,7 +265,7 @@ def test_add_variant_to_order_edits_line_for_existing_variant(
 
 
 def test_add_variant_to_order_not_allocates_stock_for_existing_variant(
-    order_with_lines, anonymous_user, anonymous_plugins, site_settings
+    order_with_lines, anonymous_plugins, site_settings
 ):
     existing_line = order_with_lines.lines.first()
     variant = existing_line.variant
@@ -284,7 +280,7 @@ def test_add_variant_to_order_not_allocates_stock_for_existing_variant(
     add_variant_to_order(
         order=order_with_lines,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,
@@ -560,7 +556,7 @@ def test_calculate_order_weight(order_with_lines):
 
 
 def test_order_weight_add_more_variant(
-    order_with_lines, anonymous_user, anonymous_plugins, site_settings
+    order_with_lines, anonymous_plugins, site_settings
 ):
     variant = order_with_lines.lines.first().variant
     line_data = OrderLineData(variant_id=str(variant.id), variant=variant, quantity=2)
@@ -568,7 +564,7 @@ def test_order_weight_add_more_variant(
     add_variant_to_order(
         order=order_with_lines,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,
@@ -583,7 +579,6 @@ def test_order_weight_add_more_variant(
 def test_order_weight_add_new_variant(
     order_with_lines,
     product,
-    anonymous_user,
     anonymous_plugins,
     site_settings,
 ):
@@ -593,7 +588,7 @@ def test_order_weight_add_new_variant(
     add_variant_to_order(
         order=order_with_lines,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,
@@ -632,7 +627,6 @@ def test_order_weight_delete_line(lines_info):
 def test_get_order_weight_non_existing_product(
     order_with_lines,
     product,
-    anonymous_user,
     anonymous_plugins,
     site_settings,
 ):
@@ -644,7 +638,7 @@ def test_get_order_weight_non_existing_product(
     add_variant_to_order(
         order=order,
         line_data=line_data,
-        user=anonymous_user,
+        user=None,
         app=None,
         manager=anonymous_plugins,
         site_settings=site_settings,